requirement.txt
streamlit
faiss-cpu
numpy
transformers
sentence-transformers
